Timer working correctly, adding doc comment
