--- src/pages/Timer/Timer.css
@@ -1,362 +1,0 @@
-.timer-page {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: calc(100vh - 60px);
-}
-
-.timer-container {
-  width: 100%;
-  max-width: 600px;
-  text-align: center;
-}
-
-.timer-header h1 {
-  font-size: 24px;
-  font-weight: normal;
-  margin-bottom: 30px;
-  color: #333;
-}
-
-.timer-display {
-  margin-bottom: 30px;
-}
-
-.small-timer {
-  font-size: 18px;
-  font-family: monospace;
-  margin-bottom: 10px;
-  color: #666;
-}
-
-.large-timer {
-  font-size: 72px;
-  font-family: monospace;
-  font-weight: bold;
-  color: #333;
-}
-
-.break-timer {
-  font-size: 72px;
-  font-family: monospace;
-  font-weight: bold;
-  color: #333;
-  margin: 20px 0;
-}
-
-.timer-progress {
-  margin-bottom: 30px;
-}
-
-.progress-bar {
-  height: 8px;
-  background-color: #eee;
-  border-radius: 4px;
-  margin-bottom: 10px;
-  overflow: hidden;
-}
-
-.progress-fill {
-  height: 100%;
-  width: 30%;
-  background-color: #333;
-  border-radius: 4px;
-  transition: width 0.5s ease;
-}
-
-.break-text {
-  font-size: 14px;
-  color: #666;
-  margin-top: 20px;
-}
-
-.break-controls {
-  display: flex;
-  justify-content: center;
-  gap: 15px;
-  margin-top: 20px;
-}
-
-.take-break-btn {
-  background-color: #333;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-  transition: background-color 0.2s;
-}
-
-.take-break-btn:hover {
-  background-color: #444;
-}
-
-.take-break-btn.disabled,
-.take-break-btn:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-}
-
-.extend-break-btn {
-  background-color: #555;
-  color: white;
-  border: none;
-  padding:
-  10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-}
-
-.extend-break-btn:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-}
-
-.end-break-btn {
-  background-color: #333;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-}
-
-.timer-controls {
-  margin-bottom: 30px;
-}
-
-.play-button {
-  background-color: transparent;
-  border: none;
-  color: #333;
-  font-size: 16px;
-  cursor: pointer;
-  padding: 10px 20px;
-}
-
-.timer-tags {
-  margin-bottom: 30px;
-}
-
-.tags-input {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  font-size: 14px;
-}
-
-.timer-actions {
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-  margin-top: 40px;
-}
-
-.save-button {
-  background-color: #666;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-  transition: background-color 0.2s;
-}
-
-.save-button:hover {
-  background-color: #555;
-}
-
-.discard-button {
-  background-color: transparent;
-  color: #333;
-  border: 1px solid #ccc;
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-  transition: all 0.2s;
-}
-
-.discard-button:hover {
-  background-color: #f5f5f5;
-}
-
-/* Add styles for the new SessionComplete component */
-.session-complete-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 80vh;
-  padding: 2rem;
-}
-
-.session-complete-card {
-  background-color: white;
-  border-radius: 16px;
-  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.1);
-  padding: 2.5rem;
-  width: 100%;
-  max-width: 600px;
-  text-align: center;
-  animation: fadeIn 0.8s ease-out;
-}
-
-@keyframes fadeIn {
-  from { opacity: 0; transform: translateY(20px); }
-  to { opacity: 1; transform: translateY(0); }
-}
-
-.session-complete-header {
-  margin-bottom: 2rem;
-  padding-bottom: 1.5rem;
-  border-bottom: 1px solid #f0f0f0;
-}
-
-.session-complete-header h1 {
-  color: #34c759;
-  font-size: 2.5rem;
-  margin-bottom: 0.5rem;
-}
-
-.session-complete-header h2 {
-  color: #555;
-  font-weight: 400;
-  font-size: 1.3rem;
-}
-
-.session-stats {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.stat-item {
-  display: flex;
-  align-items: center;
-  padding: 1rem;
-  background-color: #f8f9fa;
-  border-radius: 12px;
-  transition: transform 0.2s ease;
-}
-
-.stat-item:hover {
-  transform: translateY(-3px);
-}
-
-.stat-icon {
-  font-size: 2rem;
-  margin-right: 1rem;
-  width: 50px;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #e9f7f2;
-  border-radius: 50%;
-  color: #34c759;
-}
-
-.stat-info {
-  text-align: left;
-  flex-grow: 1;
-}
-
-.stat-info h3 {
-  font-size: 0.9rem;
-  color: #777;
-  margin-bottom: 0.4rem;
-  font-weight: 500;
-}
-
-.stat-value {
-  font-size: 1.4rem;
-  font-weight: 600;
-  color: #333;
-  margin: 0;
-}
-
-.completion-message {
-  background-color: #f6fcf8;
-  padding: 1.5rem;
-  border-radius: 10px;
-  margin-top: 1rem;
-  border-left: 4px solid #34c759;
-}
-
-.completion-message p {
-  color: #555;
-  font-size: 1.1rem;
-  margin: 0;
-  line-height: 1.5;
-}
-
-/* Return to Home button styles */
-.return-home-container {
-  margin-top: 2rem;
-}
-
-.return-home-button {
-  background-color: #34c759;
-  color: white;
-  border: none;
-  border-radius: 8px;
-  padding: 14px 32px;
-  font-size: 1.1rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  box-shadow: 0 4px 12px rgba(52, 199, 89, 0.2);
-}
-
-.return-home-button:hover {
-  background-color: #2db14f;
-  transform: translateY(-2px);
-  box-shadow: 0 6px 16px rgba(52, 199, 89, 0.3);
-}
-
-.return-home-button:active {
-  transform: translateY(0);
-  box-shadow: 0 4px 8px rgba(52, 199, 89, 0.2);
-}
-
-/* Responsive adjustments */
-@media (min-width: 768px) {
-  .session-stats {
-    flex-direction: row;
-    gap: 1rem;
-  }
-  
-  .stat-item {
-    flex: 1;
-    flex-direction: column;
-    text-align: center;
-  }
-  
-  .stat-icon {
-    margin-right: 0;
-    margin-bottom: 1rem;
-  }
-  
-  .stat-info {
-    text-align: center;
-  }
-}
-
-/* Tree type info display */
-.tree-type-info {
-  margin: 10px 0;
-  text-align: center;
-  font-size: 14px;
-  color: #666;
-}
-
-.tree-type-label {
-  font-weight: bold;
-  margin-right: 5px;
-}
